+++ packages/eslint-plugin-qwik/index.ts
@@ -1,5 +1,5 @@
 import { validLexicalScope } from './src/validLexicalScope';
-import { useMethodUsage, useMethodUsageExamples } from './src/useMethodUsage';
+import { useMethodUsage } from './src/useMethodUsage';
 import { loaderLocation } from './src/loaderLocation';
 import { noReactProps } from './src/noReactProps';
 import { preferClasslist } from './src/preferClasslist';
@@ -7,20 +7,6 @@ import { jsxNoScriptUrl } from './src/jsxNoScriptUrl';
 import { jsxKey } from './src/jsxKey';
 import { unusedServer } from './src/unusedServer';
 import { jsxImg } from './src/jsxImg';
-
-export type QwikEslintExample = {
-  code: string;
-  codeHighlight?: string;
-  description?: string;
-};
-
-export type QwikEslintExamples = Record<
-  string,
-  {
-    good: QwikEslintExample[];
-    bad: QwikEslintExample[];
-  }
->;
 
 export const rules = {
   'use-method-usage': useMethodUsage,
@@ -63,16 +49,4 @@ export const configs = {
       'qwik/jsx-img': 'error',
     },
   },
-};
-
-export const examples = {
-  'use-method-usage': useMethodUsageExamples,
-  'valid-lexical-scope': null,
-  'loader-location': null,
-  'no-react-props': null,
-  'prefer-classlist': null,
-  'jsx-no-script-url': null,
-  'jsx-key': null,
-  'unused-server': null,
-  'jsx-img': null,
 };
